Revert "merge all commit to my forked repo (#1)"
This reverts commit 1de128fc40c7e4190279ba7394e8437c220c7e91.

+++ hbase-server/src/test/java/org/apache/hadoop/hbase/client/TestSnapshotTemporaryDirectory.java
@@ -29,7 +29,6 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import java.util.UUID;
-import java.util.regex.Pattern;
 import org.apache.commons.io.FileUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
@@ -100,11 +99,10 @@ public class TestSnapshotTemporaryDirectory {
    *
    * @throws Exception on failure
    */
-  @BeforeClass
-  public static void setupCluster() throws Exception {
+  @BeforeClass public static void setupCluster() throws Exception {
     setupConf(UTIL.getConfiguration());
     UTIL.startMiniCluster(NUM_RS);
-    admin = UTIL.getAdmin();
+    admin = UTIL.getHBaseAdmin();
   }
 
   private static void setupConf(Configuration conf) {
@@ -122,11 +120,10 @@ public class TestSnapshotTemporaryDirectory {
     conf.setBoolean(SnapshotManager.HBASE_SNAPSHOT_ENABLED, true);
     conf.set(HConstants.HBASE_REGION_SPLIT_POLICY_KEY,
         ConstantSizeRegionSplitPolicy.class.getName());
-    conf.set(SnapshotDescriptionUtils.SNAPSHOT_WORKING_DIR, "file://" + new Path(TEMP_DIR, ".tmpDir").toUri());
+    conf.set(SnapshotDescriptionUtils.SNAPSHOT_WORKING_DIR, "file://" + TEMP_DIR + "/.tmpdir/");
   }
 
-  @Before
-  public void setup() throws Exception {
+  @Before public void setup() throws Exception {
     HTableDescriptor htd = new HTableDescriptor(TABLE_NAME);
     htd.setRegionReplication(getNumReplicas());
     UTIL.createTable(htd, new byte[][] { TEST_FAM }, UTIL.getConfiguration());
@@ -136,15 +133,13 @@ public class TestSnapshotTemporaryDirectory {
     return 1;
   }
 
-  @After
-  public void tearDown() throws Exception {
+  @After public void tearDown() throws Exception {
     UTIL.deleteTable(TABLE_NAME);
-    SnapshotTestingUtils.deleteAllSnapshots(UTIL.getAdmin());
+    SnapshotTestingUtils.deleteAllSnapshots(UTIL.getHBaseAdmin());
     SnapshotTestingUtils.deleteArchiveDirectory(UTIL);
   }
 
-  @AfterClass
-  public static void cleanupTest() {
+  @AfterClass public static void cleanupTest() {
     try {
       UTIL.shutdownMiniCluster();
       FileUtils.deleteDirectory(new File(TEMP_DIR));
@@ -153,14 +148,13 @@ public class TestSnapshotTemporaryDirectory {
     }
   }
 
-  @Test
-  public void testRestoreDisabledSnapshot()
+  @Test(timeout = 180000) public void testRestoreDisabledSnapshot()
       throws IOException, InterruptedException {
     long tid = System.currentTimeMillis();
     TableName tableName = TableName.valueOf("testtb-" + tid);
-    String emptySnapshot = "emptySnaptb-" + tid;
-    String snapshotName0 = "snaptb0-" + tid;
-    String snapshotName1 = "snaptb1-" + tid;
+    byte[] emptySnapshot = Bytes.toBytes("emptySnaptb-" + tid);
+    byte[] snapshotName0 = Bytes.toBytes("snaptb0-" + tid);
+    byte[] snapshotName1 = Bytes.toBytes("snaptb1-" + tid);
     int snapshot0Rows;
     int snapshot1Rows;
 
@@ -169,7 +163,7 @@ public class TestSnapshotTemporaryDirectory {
     admin.disableTable(tableName);
 
     // take an empty snapshot
-    takeSnapshot(tableName, emptySnapshot, true);
+    takeSnapshot(tableName, Bytes.toString(emptySnapshot), true);
 
     // enable table and insert data
     admin.enableTable(tableName);
@@ -180,7 +174,7 @@ public class TestSnapshotTemporaryDirectory {
     admin.disableTable(tableName);
 
     // take a snapshot
-    takeSnapshot(tableName, snapshotName0, true);
+    takeSnapshot(tableName, Bytes.toString(snapshotName0), true);
 
     // enable table and insert more data
     admin.enableTable(tableName);
@@ -191,7 +185,7 @@ public class TestSnapshotTemporaryDirectory {
 
     SnapshotTestingUtils.verifyRowCount(UTIL, tableName, snapshot1Rows);
     admin.disableTable(tableName);
-    takeSnapshot(tableName, snapshotName1, true);
+    takeSnapshot(tableName, Bytes.toString(snapshotName1), true);
 
     // Restore from snapshot-0
     admin.restoreSnapshot(snapshotName0);
@@ -220,14 +214,13 @@ public class TestSnapshotTemporaryDirectory {
     SnapshotTestingUtils.verifyReplicasCameOnline(tableName, admin, getNumReplicas());
   }
 
-  @Test
-  public void testRestoreEnabledSnapshot()
+  @Test(timeout = 180000) public void testRestoreEnabledSnapshot()
       throws IOException, InterruptedException {
     long tid = System.currentTimeMillis();
     TableName tableName = TableName.valueOf("testtb-" + tid);
-    String emptySnapshot = "emptySnaptb-" + tid;
-    String snapshotName0 = "snaptb0-" + tid;
-    String snapshotName1 = "snaptb1-" + tid;
+    byte[] emptySnapshot = Bytes.toBytes("emptySnaptb-" + tid);
+    byte[] snapshotName0 = Bytes.toBytes("snaptb0-" + tid);
+    byte[] snapshotName1 = Bytes.toBytes("snaptb1-" + tid);
     int snapshot0Rows;
     int snapshot1Rows;
 
@@ -235,7 +228,7 @@ public class TestSnapshotTemporaryDirectory {
     SnapshotTestingUtils.createTable(UTIL, tableName, getNumReplicas(), TEST_FAM);
 
     // take an empty snapshot
-    takeSnapshot(tableName, emptySnapshot, false);
+    takeSnapshot(tableName, Bytes.toString(emptySnapshot), false);
 
     // Insert data
     SnapshotTestingUtils.loadData(UTIL, tableName, 500, TEST_FAM);
@@ -244,7 +237,7 @@ public class TestSnapshotTemporaryDirectory {
     }
 
     // take a snapshot
-    takeSnapshot(tableName, snapshotName0, false);
+    takeSnapshot(tableName, Bytes.toString(snapshotName0), false);
 
     // Insert more data
     SnapshotTestingUtils.loadData(UTIL, tableName, 500, TEST_FAM);
@@ -253,7 +246,7 @@ public class TestSnapshotTemporaryDirectory {
     }
 
     SnapshotTestingUtils.verifyRowCount(UTIL, tableName, snapshot1Rows);
-    takeSnapshot(tableName, snapshotName1, false);
+    takeSnapshot(tableName, Bytes.toString(snapshotName1), false);
 
     // Restore from snapshot-0
     admin.disableTable(tableName);
@@ -288,9 +281,8 @@ public class TestSnapshotTemporaryDirectory {
    *
    * @throws Exception if snapshot does not complete successfully
    */
-  @Test
-  public void testOfflineTableSnapshot() throws Exception {
-    Admin admin = UTIL.getAdmin();
+  @Test(timeout = 300000) public void testOfflineTableSnapshot() throws Exception {
+    Admin admin = UTIL.getHBaseAdmin();
     // make sure we don't fail on listing snapshots
     SnapshotTestingUtils.assertNoSnapshots(admin);
 
@@ -313,7 +305,7 @@ public class TestSnapshotTemporaryDirectory {
 
     // take a snapshot of the disabled table
     final String SNAPSHOT_NAME = "offlineTableSnapshot";
-    String snapshot = SNAPSHOT_NAME;
+    byte[] snapshot = Bytes.toBytes(SNAPSHOT_NAME);
     takeSnapshot(TABLE_NAME, SNAPSHOT_NAME, true);
     LOG.debug("Snapshot completed.");
 
@@ -343,8 +335,7 @@ public class TestSnapshotTemporaryDirectory {
    *
    * @throws Exception if snapshot does not complete successfully
    */
-  @Test
-  public void testSnapshotCloneContents() throws Exception {
+  @Test(timeout = 180000) public void testSnapshotCloneContents() throws Exception {
     // make sure we don't fail on listing snapshots
     SnapshotTestingUtils.assertNoSnapshots(admin);
 
@@ -380,8 +371,7 @@ public class TestSnapshotTemporaryDirectory {
     admin.close();
   }
 
-  @Test
-  public void testOfflineTableSnapshotWithEmptyRegion() throws Exception {
+  @Test(timeout = 180000) public void testOfflineTableSnapshotWithEmptyRegion() throws Exception {
     // test with an empty table with one region
 
     // make sure we don't fail on listing snapshots
@@ -399,8 +389,8 @@ public class TestSnapshotTemporaryDirectory {
             LOG);
 
     // take a snapshot of the disabled table
-    String snapshot = "testOfflineTableSnapshotWithEmptyRegion";
-    takeSnapshot(TABLE_NAME, snapshot, true);
+    byte[] snapshot = Bytes.toBytes("testOfflineTableSnapshotWithEmptyRegion");
+    takeSnapshot(TABLE_NAME, Bytes.toString(snapshot), true);
     LOG.debug("Snapshot completed.");
 
     // make sure we have the snapshot
@@ -426,11 +416,12 @@ public class TestSnapshotTemporaryDirectory {
   }
 
   // Ensures that the snapshot is transferred to the proper completed snapshot directory
-  @Test
-  public void testEnsureTemporaryDirectoryTransfer() throws Exception {
-    Admin admin = UTIL.getAdmin();
+  @Test(timeout = 180000) public void testEnsureTemporaryDirectoryTransfer() throws Exception {
+    Admin admin = null;
     TableName tableName2 = TableName.valueOf("testListTableSnapshots");
     try {
+      admin = UTIL.getHBaseAdmin();
+
       HTableDescriptor htd = new HTableDescriptor(tableName2);
       UTIL.createTable(htd, new byte[][] { TEST_FAM }, UTIL.getConfiguration());
 
@@ -446,8 +437,7 @@ public class TestSnapshotTemporaryDirectory {
       takeSnapshot(TABLE_NAME, table2Snapshot1, false);
       LOG.debug("Table2Snapshot1 completed.");
 
-      List<SnapshotDescription> listTableSnapshots =
-        admin.listTableSnapshots(Pattern.compile("test.*"), Pattern.compile(".*"));
+      List<SnapshotDescription> listTableSnapshots = admin.listTableSnapshots("test.*", ".*");
       List<String> listTableSnapshotNames = new ArrayList<String>();
       assertEquals(3, listTableSnapshots.size());
       for (SnapshotDescription s : listTableSnapshots) {
@@ -457,22 +447,25 @@ public class TestSnapshotTemporaryDirectory {
       assertTrue(listTableSnapshotNames.contains(table1Snapshot2));
       assertTrue(listTableSnapshotNames.contains(table2Snapshot1));
     } finally {
-      try {
-        admin.deleteSnapshots(Pattern.compile("Table.*"));
-      } catch (SnapshotDoesNotExistException ignore) {
+      if (admin != null) {
+        try {
+          admin.deleteSnapshots("Table.*");
+        } catch (SnapshotDoesNotExistException ignore) {
+        }
+        if (admin.tableExists(tableName2)) {
+          UTIL.deleteTable(tableName2);
+        }
+        admin.close();
       }
-      if (admin.tableExists(tableName2)) {
-        UTIL.deleteTable(tableName2);
-      }
-      admin.close();
     }
   }
 
   private void takeSnapshot(TableName tableName, String snapshotName, boolean disabled)
       throws IOException {
     SnapshotType type = disabled ? SnapshotType.DISABLED : SnapshotType.FLUSH;
-    SnapshotDescription desc = new SnapshotDescription(snapshotName, tableName, type, null, -1,
-        manifestVersion, null);
+    SnapshotDescription desc =
+        new SnapshotDescription(snapshotName, tableName.getNameAsString(), type, null, -1,
+            manifestVersion);
     admin.snapshot(desc);
   }
 }
